uint64_t fix

+++ src/libponyrt/ds/fun.h
@@ -22,7 +22,7 @@ typedef void (*free_size_fn)(size_t size, void* data);
 
 PONY_API uint64_t ponyint_hash_block(const void* p, size_t len);
 
-uint64_t ponyint_hash_str(const char* str);
+size_t ponyint_hash_str(const char* str);
 
 size_t ponyint_hash_ptr(const void* p);
